--- a/lib/quick-sort.js
+++ b/lib/quick-sort.js
@@ -122,11 +122,11 @@ function cloneSort(comparator) {
  */
 
 let sortCache = new WeakMap();  // Cache for specialized sorts.
-exports.quickSort = function (ary, comparator) {
+exports.quickSort = function (ary, comparator, start = 0) {
   let doQuickSort = sortCache.get(comparator);
   if (doQuickSort === void 0) {
     doQuickSort = cloneSort(comparator);
     sortCache.set(comparator, doQuickSort);
   }
-  doQuickSort(ary, comparator, 0, ary.length - 1);
+  doQuickSort(ary, comparator, start, ary.length - 1);
 };