--- a/lib/source-map-consumer.js
+++ b/lib/source-map-consumer.js
@@ -556,7 +556,16 @@ BasicSourceMapConsumer.prototype._parseMappings =
 
         generatedMappings.push(mapping);
         if (typeof mapping.originalLine === 'number') {
-          originalMappings.push(mapping);
+          // This code used to just do: originalMappings.push(mapping).
+          // Now it sorts original mappings already by source during parsing.
+          let currentSource = mapping.source;
+          while (originalMappings.length <= currentSource) {
+            originalMappings.push(null);
+          }
+          if (originalMappings[currentSource] === null) {
+            originalMappings[currentSource] = [];
+          }
+          originalMappings[currentSource].push(mapping);
         }
       }
     }
@@ -564,8 +573,14 @@ BasicSourceMapConsumer.prototype._parseMappings =
     quickSort(generatedMappings, util.compareByGeneratedPositionsDeflated);
     this.__generatedMappings = generatedMappings;
 
-    quickSort(originalMappings, util.compareByOriginalPositions);
-    this.__originalMappings = originalMappings;
+    // The code used to sort the whole array:
+    // quickSort(originalMappings, util.compareByOriginalPositions);
+    for (var i = 0; i < originalMappings.length; i++) {
+      if (originalMappings[i] != null) {
+        quickSort(originalMappings[i], util.compareByOriginalPositionsNoSource);
+      }
+    }
+    this.__originalMappings = [].concat(...originalMappings);
   };
 
 /**