--- a/lib/source-map-consumer.js
+++ b/lib/source-map-consumer.js
@@ -493,11 +493,6 @@ BasicSourceMapConsumer.prototype._parseMappings =
         mapping = new Mapping();
         mapping.generatedLine = generatedLine;
 
-        // Because each offset is encoded relative to the previous one,
-        // many segments often have the same encoding. We can exploit this
-        // fact by caching the parsed variable length fields of each segment,
-        // allowing us to avoid a second parse if we encounter the same
-        // segment again.
         for (end = index; end < length; end++) {
           if (this._charIsMappingSeparator(aStr, end)) {
             break;
@@ -505,27 +500,20 @@ BasicSourceMapConsumer.prototype._parseMappings =
         }
         str = aStr.slice(index, end);
 
-        segment = cachedSegments[str];
-        if (segment) {
-          index += str.length;
-        } else {
-          segment = [];
-          while (index < end) {
-            base64VLQ.decode(aStr, index, temp);
-            value = temp.value;
-            index = temp.rest;
-            segment.push(value);
-          }
+        segment = [];
+        while (index < end) {
+          base64VLQ.decode(aStr, index, temp);
+          value = temp.value;
+          index = temp.rest;
+          segment.push(value);
+        }
 
-          if (segment.length === 2) {
-            throw new Error('Found a source, but no line and column');
-          }
+        if (segment.length === 2) {
+          throw new Error('Found a source, but no line and column');
+        }
 
-          if (segment.length === 3) {
-            throw new Error('Found a source and line, but no column');
-          }
-
-          cachedSegments[str] = segment;
+        if (segment.length === 3) {
+          throw new Error('Found a source and line, but no column');
         }
 
         // Generated column.