--- a/lib/source-map-consumer.js
+++ b/lib/source-map-consumer.js
@@ -464,6 +464,36 @@ function Mapping() {
  * query (the ordered arrays in the `this.__generatedMappings` and
  * `this.__originalMappings` properties).
  */
+
+const compareGenerated = util.compareByGeneratedPositionsDeflatedNoLine;
+function sortGenerated(array, start) {
+  let l = array.length;
+  let n = array.length - start;
+  if (n <= 1) {
+    return;
+  } else if (n == 2) {
+    let a = array[start];
+    let b = array[start + 1];
+    if (compareGenerated(a, b) > 0) {
+      array[start] = b;
+      array[start + 1] = a;
+    }
+  } else if (n < 20) {
+    for (let i = start; i < l; i++) {
+      for (let j = i; j > start; j--) {
+        let a = array[j - 1];
+        let b = array[j];
+        if (compareGenerated(a, b) <= 0) {
+          break;
+        }
+        array[j - 1] = b;
+        array[j] = a;
+      }
+    }
+  } else {
+    quickSort(array, compareGenerated, start);
+  }
+}
 BasicSourceMapConsumer.prototype._parseMappings =
   function SourceMapConsumer_parseMappings(aStr, aSourceRoot) {
     var generatedLine = 1;
@@ -480,11 +510,15 @@ BasicSourceMapConsumer.prototype._parseMappings =
     var generatedMappings = [];
     var mapping, str, segment, end, value;
 
+    let subarrayStart = 0;
     while (index < length) {
       if (aStr.charAt(index) === ';') {
         generatedLine++;
         index++;
         previousGeneratedColumn = 0;
+
+        sortGenerated(generatedMappings, subarrayStart);
+        subarrayStart = generatedMappings.length;
       }
       else if (aStr.charAt(index) === ',') {
         index++;
@@ -558,7 +592,7 @@ BasicSourceMapConsumer.prototype._parseMappings =
       }
     }
 
-    quickSort(generatedMappings, util.compareByGeneratedPositionsDeflated);
+    sortGenerated(generatedMappings, subarrayStart);
     this.__generatedMappings = generatedMappings;
 
     // The code used to sort the whole array: