--- a/lib/util.js
+++ b/lib/util.js
@@ -386,11 +386,6 @@ function compareByOriginalPositions(mappingA, mappingB, onlyCompareOriginal) {
 exports.compareByOriginalPositions = compareByOriginalPositions;
 
 function compareByOriginalPositionsNoSource(mappingA, mappingB, onlyCompareOriginal) {
-  // var cmp = strcmp(mappingA.source, mappingB.source);
-  // if (cmp !== 0) {
-  //   return cmp;
-  // }
-
   var cmp
 
   cmp = mappingA.originalLine - mappingB.originalLine;