@7rulnik 2024-03-21
Applied optimizations:
* Optimizing Sorting - Argument Adaptation
* Optimizing Sorting - Monomorphisation
* Optimizing Parsing - Removing Segment Cache
* Optimizing Sorting - Algorithmic Improvements


TO DO:
* Optimizing Parsing - Reducing GC Pressure
* Optimizing Parsing - Using Uint8Array Instead of a String.